feat(util): rename jsdom setup file
