+++ jest/setup/index.js
@@ -5,4 +5,4 @@ import '@testing-library/jest-dom/extend-expect.js'
 
 import './match-media.js'
 import './react-dom.js'
-import './jsdom.js'
+import './inner-text.js'
